pyproject.toml
[project]
name = "trachhub"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "flask[async]",
    "bleak<0.22.0",
    "hypercorn",
    "sqlalchemy",
    "websockets",
    "python-dotenv",
    "matplotlib",
    "numpy",
    "psycopg2",
]

[dependency-groups]
dev = [
    "bless",
]
